Add much needed update and delete functionality to the api controllers

+++ OMS-API/Services/ItemService.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.EntityFrameworkCore;
@@ -18,13 +19,24 @@ namespace OMSAPI.Services
         {
             var item = _context.Items.FirstOrDefault( it => it.Id == itemId );
             if(item != null) {
+                if (itemPresentOnSalesOrders(item)) {
+                    return new DatabaseOperationStatus {
+                        StatusOk = false,
+                        Message = $"Cannot remove the item because it is used in at least one sales order"
+                    };
+                }
                 _context.Items.Remove(item);
                 return SaveChanges();
             }
             return new DatabaseOperationStatus {
-                StatusOk = true,
+                StatusOk = false,
                 Message = $"There is no item with id { itemId }"
             };
+        }
+
+        private bool itemPresentOnSalesOrders(Item item)
+        {
+            return _context.SalesOrderLines.Where(line => line.ItemId == item.Id).Count() > 0;
         }
 
         public Item Get(int itemId)
@@ -45,7 +57,9 @@ namespace OMSAPI.Services
                 return Modify(tracked);
             }
             _context.Items.Add(item);
-            return SaveChanges();
+            var status = SaveChanges();
+            status.NewRecordId = item.Id;
+            return status;
         }
 
         public DatabaseOperationStatus Modify(Item item)
